Move Photoshopb class to internal namespace
