+++ src/pngimage.cpp
@@ -13,11 +13,13 @@
 #include "image.hpp"
 #include "image_int.hpp"
 #include "jpgimage.hpp"
+#include "photoshop.hpp"
 #include "pngchunk_int.hpp"
 #include "pngimage.hpp"
 #include "tiffimage.hpp"
 #include "types.hpp"
 
+#include <array>
 #include <iostream>
 
 namespace {
